% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPredictors.R
\name{buildPredictors}
\alias{buildPredictors}
\title{Build Predictors}
\usage{
buildPredictors(x)
}
\arguments{
\item{x}{An xts object containing stock quotes (OHLCV info) from quant mod}
}
\value{
xts object of predictors (not lagged) to be used to predict the trade signal
}
\description{
takes and xts object and uses TTR package to create non lagged formated set of predictors (typically used to predict the trade Flag T.flag())
}
\author{
Matthew Davis
}

