% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStockDataSetFromEnv.R
\name{buildStockDataSetFromEnv}
\alias{buildStockDataSetFromEnv}
\title{Build Stock Data Predictors from Enviroment}
\usage{
buildStockDataSetFromEnv(e, stockNameList = NULL, verbose = FALSE)
}
\arguments{
\item{e}{enviroment where the xts stock data is stored}

\item{stockNameList}{a list of stockNames}

\item{verbose}{print debugging output}
}
\value{
a data frame of merge stock data
}
\description{
takes and enviroment with xts stock data, engineers features and merges it into one data frame
}
