% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOutliersByQuantile.R
\name{findOutliersByQuantile}
\alias{findOutliersByQuantile}
\title{Find Outliers by quantile}
\usage{
findOutliersByQuantile(mydata, r = 1.8, verbose = FALSE)
}
\arguments{
\item{mydata}{a data frame}

\item{r}{IQR range mulitplier}
}
\value{
logical vector of outliers
}
\description{
takes a data frame and returns an index with IRQ outliers
}
