% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixColnames.R
\name{fixColnames}
\alias{fixColnames}
\title{fix column names}
\usage{
fixColnames(e, stockNameList = NULL)
}
\arguments{
\item{e}{and enviroment where stock xts data is stored}

\item{stockNameList}{list of stocks to correct names for, ls(e is used if null)}
}
\description{
changes xts object column names from ex: GE.Open to Open
}
