% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXtsFeatures.R
\name{getXtsFeatures}
\alias{getXtsFeatures}
\title{Build Predictors}
\usage{
getXtsFeatures(x, wid = 20, verbose = FALSE, stockName = NULL,
  clusterModel = NULL)
}
\arguments{
\item{x}{An xts object containing stock quotes (OHLCV info) from quant mod}

\item{funList}{a list of functions for to build features from and xts object}

\item{funNames}{column names for the the output xts}
}
\value{
and xts with the funNames and ouput from funList

xts object of predictors (not lagged) to be used to predict the trade signal
}
\description{
takes and xts object and uses TTR package to create non lagged formated set of predictors (typically used to predict the trade Flag T.flag())
}
\author{
Matthew Davis
}
