% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfeatures.R
\name{myMACD}
\alias{myMACD}
\title{myMACD}
\usage{
myMACD(x)
}
\arguments{
\item{x}{and xts object from quantmod}
}
\value{
and xts object
}
\description{
functions to build predictors varaibles from an xts quantmod object
}
