% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500Symbols.R
\docType{data}
\name{sp500Symbols}
\alias{sp500Symbols}
\title{a vector of the sp500 symbols}
\format{An object of class \code{character} of length 591.}
\usage{
sp500Symbols
}
\description{
a vector of the sp500 symbols
}
\keyword{datasets}
