% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockData.R
\docType{data}
\name{stockData}
\alias{stockData}
\title{merged xts of stock Open High Low Close Volume data starting Jan 31 1999}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 4711 rows and 120 columns.}
\usage{
stockData
}
\description{
merged xts of stock Open High Low Close Volume data starting Jan 31 1999
}
\keyword{datasets}
