% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tind.R
\name{T.ind}
\alias{T.ind}
\title{Trade Signal Detector}
\usage{
T.ind(quotes, tgt.margin = 0.025, n.days = 10)
}
\arguments{
\item{quotes}{An xts object containing stock quotes (build from getSymbols in quantmod)}

\item{tgt.margin}{percent increase or decrease in high or low price}

\item{n.days}{Number of days ahead}
}
\value{
xts object of trade signals, buy when T.ind > 0.1, Sell < -0.1, else hold.
}
\description{
creates a buy sell hold trade signal from stock quotes downloaded using quatmod
}
\author{
Matthew Davis
}
