% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickerEnvUpdate.R
\name{tickerEnvUpdate}
\alias{tickerEnvUpdate}
\title{Ticker Enviroment Update}
\usage{
tickerEnvUpdate(e, symbolList = NULL, startDate = "2000-01-03",
  endDate = NULL)
}
\arguments{
\item{e}{enviroment to update}

\item{symbolList}{char vector of symbols to look up}

\item{startDate}{character start date}

\item{endDate}{character end date (optional)}
}
\value{
a list of stock tickers that failed to update
}
\description{
updates and enviroment with ticker symbols
}
