% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade simulator.R
\name{tradeSimulator}
\alias{tradeSimulator}
\title{Trade Simulator}
\usage{
tradeSimulator(model, e, params, verbose = TRUE)
}
\arguments{
\item{e}{an enviroment with xts OHLC data}

\item{params}{a list of parameters}

\item{preds}{an txs object of buy sell hold probablities}
}
