% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateXts.R
\name{updateXts}
\alias{updateXts}
\title{Update my merged List of stock tickers}
\usage{
updateXts(mydata)
}
\arguments{
\item{symbol.list}{a merged xts object where the columns are stock OHLC Data}
}
\value{
merged xts of stock tickers updated to the previous day.
}
\description{
looks up a list of stock tickers and returns a merged xts of them
}
\author{
Matthew Davis
}
