% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tflag.R
\name{T.flag}
\alias{T.flag}
\title{Trade Signal Flag}
\usage{
T.flag(x, sig = 0.1)
}
\arguments{
\item{x}{An xts object containing output from T.ind()}

\item{sig}{Signal strength that triggers a buy or sell}
}
\value{
xts object of trade signals,buy = 1, sell = -1, hold = 0
}
\description{
flags when to buy sell and hold a stock, after inputing trade signal from T.ind()
}
\author{
Matthew Davis
}
